add bookmark purpose to rubric association
this commit adds the bookmark purpose to the rubric association. this
field was missing when creating a rubric association for a course or
account. it was causing the rubrics to be deleted when the assignment
removed the rubric association if there was only one rubric association.

closes EVAL-5209
flag=enhanced_rubrics

test plan:
- create a rubric assignment for a course and account
- go to an assignment and attach the rubric to the assignment
- remove the rubric from the assignment
- verify the rubric is not deleted
- create a rubric from the assignments page (assignment rubric)
- remove the rubric from the assignment
- verify that the rubric is deleted

Change-Id: I61c87cee95726aaf115e644c5ff7de8c9ae5e035
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/385382
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
QA-Review: Chris Soto <christopher.soto@instructure.com>
Product-Review: Yasi Afsharnive <yasaman.afsharnive@instructure.com>