Check all submission attempts for a report
When there is a later submission that has an identical attachment
and we already have a report for that attachment, then when we generate
a url with an attempt query to find a report for that specific attempt
it does not find the report we already have because the submission
time does not match. The report can be found if we look for the report
without using the attempt number.

closes INTEROP-9429
flag=none

Test plan:
This test is done in a 'prod' console to test directly on the
offending situation as testing locally would prove much more difficult.
(we'll be using the last "bad" example from the ticket)
- ssh iad-beta, get rails console, switch shard, get relevant variables
  from the details of the example in the ticket
- cd Submission (so that 'self' is Submission)
TEST OLD CODE:
- copy and paste the originality_report_url method into the console
  putting it into the Submission class
- run `sub.originality_report_url(asset_string, user, attempt)`
  - you should get `nil` returned
- run it again without the `attempt` parameter
  - you should get a url that doesn't have an "attempt" query
TEST NEW CODE:
- copy and paste the new originality_report_url method into the console
  putting it into the Submission class
- you will also need to individually paste in each of the three private
  helper methods that go with it
- run `sub.originality_report_url(asset_string, user, attempt)`
  - you should get a url that doesn't have an "attempt" query
- run it again without the `attempt` parameter
  - you should still get a url that doesn't have an "attempt" query

Change-Id: I395d59b0f8ab745b2ef09e7454a9bbb0456757be
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386183
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: David Varga <d.varga@instructure.com>
Reviewed-by: David Varga <d.varga@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>