Add API to get an account's brand config JSON
Adds an API that is very similar to the existing brand config API
which fetches the config for the domain root account, but this one
allows specifying an account or course. It similarly redirects to the
appropriate file on the CDN. Because this API is not needed to power
a login page, it requires the user to be logged in. It also does not
open CORS like the existing API.

closes CLX-2698
flag = none

Test plan:
 - Enable sub-account branding on the root account's settings page
 - Create and apply a theme on the root account
 - Create and apply a different theme on a subaccount
 - GET /api/v1/accounts/<root account>/brand_variables
 - Expect the response to match the root account's theme
 - GET /api/v1/accounts/<subaccount>/brand_variables
 - Expect the response to match the subaccount's theme
 - Create a course in the subaccount
 - GET /api/v1/courses/<course>/brand_variables
 - Expect the response to match the course's subaccount's theme

Change-Id: Ib6a74bd9296d8f64ffc90d15a126efb87edf11bb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386514
Reviewed-by: Ildiko Gera <ildiko.gera@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>