Add "Ignite AI" screenreader text to Smart Search heading
Fixes FOO-5638
flag=smart_search_enhanced_ui

Accessibility audit caught that the "Ignite AI" Icon on the
page title was completely "invisible" to screenreaders due
to the presence of an aria-hidden attribute. The ticket
calls for that to be removed and alt text added, but the
InstUI Icon components do not provide this functionality,
so I left the icon the way it was and instead added an
adjacent <ScreenReaderContent> with the "Ignite AI" text.

Test plan:
* Turn on your screenreader and navigate to the Smart Search
  page for a course
* Get the screenreader to read the title heading
* It should voice out:
        "Heading level 1, Ignite AI, Smart Search"

Change-Id: I5365f937447105175d417be61d516be83097e299
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386643
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>