spec: Implement a mock LTI tool for Selenium testing
This adds a few things for testing an LTI tool launch in Canvas,
as well as a Selenium test that does the launch.

First, in spec/selenium/test_setup/test_only_routes.rb, there is now
a place to define endpoints that will only be available in Selenium
tests.

The endpoints defined in there mimic the endpoints that an LTI 1.3
tool would define, in order to get a basic tool launch happening in
Canvas.

The /test/mock_lti/ui endpoints shows an HTML page with a single button,
which will send a request to subscribe to the notification service when
clicked.

A Selenium test using these new endpoints is also part of this commit.

[skip-crystalball]

test plan:
- specs pass

flag = none

fixes INTEROP-9239

Change-Id: Ic184aeec0b48f4bcd941172fac22f76910eaa4cc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386938
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: David Varga <d.varga@instructure.com>
QA-Review: David Varga <d.varga@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>