Add Max Expiration to Access Tokens
This commit adds a maximum expiration time of 120 days to User-generated
access tokens for students.

A "student" is defined as a user who does not have an Admin role in the
account, and whose enrollments are all of type "StudentEnrollment".

Also included in this commit is the "required" field for the token
purpose. A purpose has always been required, but it wasn't shown in the
UI.

test plan:
- Log in as a user who has no account admin role, and has at least one
  enrollment in any course as a student, and no enrollments with roles
  other than student.
- Go to the Access Tokens section in the user profile settings page.
- Click "Create New Access Token"
- Verify that the "Purpose" field is marked as required (star character)
- Verify that the "Expiration" field has a maximum value of 120 days

- Turn off the `student_access_token_forced_expiration` feature flag
- Refresh the page
- Ensure that the student user can create a token with no expiration.

closes INTEROP-9526 INTEROP-9525
flag=student_access_token_forced_expiration

Change-Id: Ib60fe1a583682af29d27dffa3c9d5e06bf0d58ff
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/388229
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>