Fix 500 error when SubAssignment Submissions are deleted
When collecting SubAssignment Submissions to be returned
in submission json, if the SubAssignment Submissions are
deleted then the query will return nil. This is also true
for Submissions that do not exist on SubAssignments. This
issue will be fixed in EGG-1916.  For now, if nil is
returned then the SubAssignment is skipped and is not
added to the json.

This PS also includes unit tests for Discussion Checkpoints
management in SubmissionLifecycleManager to document how
submissions are managed.

fixes EGG-1785

flag=discussion_checkpoints

Test plan:
- Tests are sufficient & Jenkins passes

Change-Id: Id25436d326a6a9e3c2dc15f946eaf42d8cdc0296
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/388638
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Caleb Guanzon <cguanzon@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Sam Garza <sam.garza@instructure.com>