Add developer key ID and vendor to client_identifiers array
This array is used when determining which characteristics about
a request we want to use to decide which rate-limiting bucket
we want to put them in.

The developer key is found by calling find_cached (this can be
tested by putting a breakpoint in that method and seeing that
it gets called while the tests are running), which means
this shouldn't be adding an extra query on top of what
we already query for.

fixes INTEROP-9369

flag = none

test plan:
- specs pass

Change-Id: Ic62d730626cc500c2394d386a9ae2b186e915b32
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/388554
Reviewed-by: Paul Gray <paul.gray@instructure.com>
QA-Review: Paul Gray <paul.gray@instructure.com>
Tested-by: James Butters <jbutters@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>