Read Inbox recipients when they are added
Currently, when we add a User to the recipients list,
we are not communicating that to the screenreader.

This causes that the user does not know whether they were
able to select the recipient successfully or not.

closes VICE-5417
flag=none

test plan:

Note: If you are using `VoiceOver`, make sure to use `Safari` and not `Chrome`. We do not officially support VoiceOver on Chrome.

- Go to the Inbox
- Enable VoiceOver
- Compose a new message
- Select a user from the Users submenu
- Click on the user
- Voiceover should read that the user has been added
- Select another user
- Press Enter
- Voiceover should read that the user has been added

Change-Id: I98fa1d8738593be48a92a9ef7e3aab4478320652
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/388117
QA-Review: Dora Csolakov <dora.csolakov@instructure.com>
Reviewed-by: Dora Csolakov <dora.csolakov@instructure.com>
Product-Review: Bence Árpási <bence.arpasi@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>