moderated grading navigation and state tracking
Adds hasProvisionalGradeByCurrentUser field to track
provisional grade completion and updates needs_grading
sort to exclude already-graded submissions, fixing
navigation for provisional graders in SpeedGrader.

flag=moderated_grading_modernized_speedgrader
closes EVAL-5499

Test Plan:
- Create a moderated grading assignment with 1 provisional grader
- Before a student submits to the assignment view it as the moderator
and provisional grader and ensure there is not an orange dot present
next to any students.
- Submit to the assignment as one student
- As the provisional grader navigate to the speedgrader without a
student_id parameter and ensure it navigates to the student that
submitted and is waiting for a grade
- Ensure there is a orange dot next to their name indicating that this
is submission needs grading.
- grade the submission and ensure that the orange dot changes to a
checkmark and that it stays a checkmark after a page refresh
- Submit as another student
- As the provisional grader navigate to the speedgrader without a
student_id parameter and ensure it navigates to the student that just
submitted and is waiting for a grade and not the previously graded
student
- Ensure that if you take away a grade that you had given previously,
the check mark goes away and the orange dot comes back, and you're able
to give a different provisional grade after that.
- Once all the grader slots have been filled, view the assignments in
SpeedGrader as a grader that did not get one of the provisional grader
slots. Ensure that there are no check marks or orange dots in the
student drop-down.
- As the moderator, navigate to the SpeedGrader assignment and ensure
that you are navigated to the first ungraded student alphabetically.
- Ensure that you are only given a check mark when the moderator enters
 a custom grade into the grade input. And if they take away that custom
 grade, ensure that the check mark goes back to an orange dot.
- If the moderator doesn't enter a custom grade and just selects a grade
 from the show details radio buttons, ensure that it stays as an orange
 dot.
- The only way a moderator should have the check mark show up is if they
 have entered their own custom provisional grade.
- Make one selection for a student from the provisional graders and
choose one custom score
- Release grades for that moderated assignment from the moderate page
- ensure all students that had grades finalized have a check mark next
 to their names for all graders including ones that did not get a slot
  on the assignment.
- Ensure you can grade with any grader and see the checkmark/orange dot
 appropriately if there is a grade present. It should behave like a
  normal assignment now.

Change-Id: I00f6959ca0ccf34ba8cf45af72e61d4c3b70042d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/389862
Reviewed-by: Daniel Kis-Pal <daniel.kispal@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Yasi Afsharnive <yasaman.afsharnive@instructure.com>